update the Calculator.app to version 0.2
