--- Under Construction/Calculator.app/Calculator
@@ -1,3 +1,0 @@
-#!/bin/sh
-HERE="$(dirname "$(readlink -f "${0}")")"
-exec "${HERE}/Resources/calculator.py" "$@"
